pyproject.toml
[tool.poetry]
name = "phidata-start"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
phidata = "^2.3.41"
openai = "^1.12.0"
duckduckgo-search = "^4.4.3"
google-cloud-aiplatform = "^1.42.1"
ollama = "^0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
